Projects/Multi-Agent Research Synthesis System/requirements.txt
# Core LangChain and LangGraph
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langchain-community==0.3.5
langchain-core==0.3.17

# Vector Store
faiss-cpu==1.9.0
# For GPU: faiss-gpu==1.9.0

# LLM APIs
openai==1.54.3
anthropic==0.39.0

# Paper Search
arxiv==2.1.3

# Utilities
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.10.3
tiktoken==0.8.0

# Optional: Monitoring and Logging
langsmith==0.1.147
tenacity==9.0.0